add mask.go
